+++ database/migrations/2021_01_07_124049_create_ref_models_table.php
@@ -15,7 +15,6 @@ class CreateRefModelsTable extends Migration
     {
         Schema::create('ref_models', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->string('domain_name')->unique();
             $table->string('class_name')->unique();
             $table->string('table_name');
             $table->timestamps();