+++ changelog.md
@@ -2,6 +2,11 @@
 
 All notable changes to `DomainReflection` will be documented in this file.
 
+## Version 2.0
+- Removed domain name from ref_models
+- Added "related_class_name" and "related_model_id" for ref_relations
+- Fixed dependencies
+
 ## Version 1.1
 - Added scopes